refactor: moving the Legacy VMSS state migration into that package
