+++ internal/services/legacy/migration/legacy_vmss_v0_to_v1.go
